+++ sdk/identity/identity/src/plugins/consumer.ts
@@ -11,7 +11,6 @@ import {
 /**
  * The context passed to an Identity plugin. This contains objects that
  * plugins can use to set backend implementations.
- * @internal
  */
 const pluginContext: AzurePluginContext = {
   cachePluginControl: msalNodeFlowCacheControl,