+++ sdk/storagediscovery/arm-storagediscovery/samples/v1/javascript/storageDiscoveryWorkspacesDeleteSample.js
@@ -5,24 +5,20 @@ const { StorageDiscoveryClient } = require("@azure/arm-storagediscovery");
 const { DefaultAzureCredential } = require("@azure/identity");
 
 /**
- * This sample demonstrates how to get a StorageDiscoveryWorkspace
+ * This sample demonstrates how to delete a StorageDiscoveryWorkspace
  *
- * @summary get a StorageDiscoveryWorkspace
- * x-ms-original-file: 2025-06-01-preview/StorageDiscoveryWorkspaces_Get.json
+ * @summary delete a StorageDiscoveryWorkspace
+ * x-ms-original-file: 2025-09-01/StorageDiscoveryWorkspaces_Delete.json
  */
-async function getAStorageDiscoveryWorkspace() {
+async function deleteAStorageDiscoveryWorkspace() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "b79cb3ba-745e-5d9a-8903-4a02327a7e09";
   const client = new StorageDiscoveryClient(credential, subscriptionId);
-  const result = await client.storageDiscoveryWorkspaces.get(
-    "sample-rg",
-    "Sample-Storage-Workspace",
-  );
-  console.log(result);
+  await client.storageDiscoveryWorkspaces.delete("sample-rg", "sampleworkspace");
 }
 
 async function main() {
-  await getAStorageDiscoveryWorkspace();
+  await deleteAStorageDiscoveryWorkspace();
 }
 
 main().catch(console.error);