+++ sdk/identity/identity/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 4.13.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 4.13.0 (2025-10-07)
 
 ### Features Added